added error handling function require in 'transfer' method in TokenManager

+++ src/main/java/com/skysoft/vaultlogic/blockchain/truffle/contracts/services/ITokenManager.sol
@@ -2,7 +2,7 @@ pragma solidity 0.4.24;
 
 interface ITokenManager {
 
-    function balanceOf(address _owner) external view returns (uint);
+    function balanceOf(address _owner) external view returns (uint256);
 
     function transfer(address _recipient, uint _value) external;
